calendar/requirements.txt
rpio.gpio
requests
pygame
pytest
unittest